Add support for custom event dispatching
### Description

This addresses #615 by adding the new ACTIVITY parameters. When this
is specified, it overrides the default event-to-activity heuristic in
#find_activity and selects the activity to run directly. This allows
the end user to write their logic in the workflows when required.

### Testing

I've added unit tests for these, but I haven't added integration or
end-to-end tests, I was unable to get the environment set up
correctly.